requirements.txt
langchain
langchain-community
langchain-openai
langchain-anthropic
langchain_mistralai

pydantic
tiktoken
python-dotenv

wikipedia
duckduckgo-search

streamlit
streamlit-chat